chore: extract types into file, rename b3-common
