+++ packages/opentelemetry-propagator-b3/src/common.ts
